User/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-PyMongo==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.1.1
pika==1.3.2
pymongo==4.3.3
requests==2.31.0
urllib3==2.0.2
Werkzeug==2.3.4
pycryptodome==3.11.0
cryptography==42.0.8